Refactored 'lds' to 'srlm'
