+++ .flaskenv
@@ -1,2 +1,2 @@
-FLASK_APP=api\lds\run.py
+FLASK_APP=api\srlm\run.py
 FLASK_DEBUG=1
