+++ api/__init__.py
@@ -1,3 +1,3 @@
-import api.lds
+import api.srlm
 
-from api.lds import asgi
+from api.srlm import asgi
