--- api/lds/__init__.py
@@ -1,7 +1,0 @@
-import api.lds.definitions
-import api.lds.app
-import api.lds.logger
-import api.lds.exceptions
-import api.lds.api_access
-
-from api.lds.run import asgi_app as asgi
