--- api/lds/api_access/__init__.py
@@ -1,1 +1,0 @@
-from api.lds.api_access import models
