--- api/lds/app/api/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint('api', __name__)
-
-from api.lds.app.api import errors, players, tokens, users, auth
