--- api/lds/logger/__init__.py
@@ -1,2 +1,0 @@
-from api.lds.logger.config import LogConfig, MailConfig
-from api.lds.logger.handlers import get_logger
