+++ api/srlm/app/api/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('api', __name__)
+
+from api.srlm.app.api import errors, players, tokens, users, auth
