+++ api/srlm/app/api/errors.py
@@ -1,6 +1,6 @@
 from werkzeug.http import HTTP_STATUS_CODES
 from werkzeug.exceptions import HTTPException
-from api.lds.app.api import bp
+from api.srlm.app.api import bp
 
 
 def error_response(status_code, message=None):
