+++ api/srlm/app/auth/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint('auth', __name__)
 
-from api.lds.app.auth import email, functions
+from api.srlm.app.auth import email, functions
