+++ api/srlm/app/templates/auth/new_user_email.html
