+++ api/srlm/app/templates/auth/reset_password_email.txt
