+++ api/srlm/definitions/__init__.py
@@ -1,5 +1,5 @@
 import os
 
-from api.lds.definitions.paths import ROOT_DIR
+from api.srlm.definitions.paths import ROOT_DIR
 
 app_name = os.getenv('SRLM_APP_NAME') or 'Slapshot: Rebound League Manager'
