+++ api/srlm/definitions/paths.py
@@ -1,7 +1,7 @@
 import os.path
 from pathlib import Path
 
-# Get directory of paths.py ([project_path]/lds/definitions/)
+# Get directory of paths.py ([project_path]/srlm/definitions/)
 path = os.path.dirname(os.path.abspath(__file__))
 # Go up 3 levels ([project_path]/)
 for _ in range(2):
