+++ api/srlm/logger/__init__.py
@@ -1,0 +1,2 @@
+from api.srlm.logger.config import LogConfig, MailConfig
+from api.srlm.logger.handlers import get_logger
