+++ setup.sh
@@ -16,3 +16,4 @@ export API_ACCESS_DB=api_access
 export MAIL_SERVER=localhost
 export MAIL_PORT=8025
 export SRLM_APP_NAME=
+export FLASK_APP=api\srlm\run.py
